requirements.txt
allure-behave==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
behave==1.2.6
contextlib2==21.6.0
parse-type==0.6.0
parse==1.19.0
pip==23.1.2
pluggy==1.0.0
pycryptodome==3.18.0
PyYAML==6.0
schema==0.7.5
setuptools==67.7.2
six==1.16.0
wheel==0.40.0
